pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-disaster_env"
version = "0.1.0"
description = "Flood disaster response simulation environment for OpenEnv — LLM agent coordinates rescue resources across flood-affected zones"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "disaster_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["disaster_env", "disaster_env.server"]
package-dir = { "disaster_env" = ".", "disaster_env.server" = "server" }